// src/types/post.ts
export interface ICreateComments {
  text: Descendant[];
}
// ...
export interface ISharePost {
  audience: Audience;
  mediaUrls: any[];
  text: Descendant[];
  editMode: boolean;
  id?: string;
  sharePostId?: string;
  postType?: string;
  sharedPostType?: string;
  location: ILocationSelect | null;
  currentPosition: any[];
}